// src/app/services/api/transaction/transaction.service.ts
export class AddInvoiceDto implements IAddInvoiceDto {
    transactionId?: number;
    fileId?: string;

    constructor(data?: IAddInvoiceDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.transactionId = _data["transactionId"];
            this.fileId = _data["fileId"];
        }
    }

    static fromJS(data: any): AddInvoiceDto {
        data = typeof data === 'object' ? data : {};
        let result = new AddInvoiceDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["transactionId"] = this.transactionId;
        data["fileId"] = this.fileId;
        return data;
    }
}

export interface IAddInvoiceDto {
    transactionId?: number;
    fileId?: string;
}

export class CurrencyDto implements ICurrencyDto {
    code?: string | undefined;
    isDefault?: boolean;
    factorToDefaultCurrency?: number;

    constructor(data?: ICurrencyDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.code = _data["code"];
            this.isDefault = _data["isDefault"];
            this.factorToDefaultCurrency = _data["factorToDefaultCurrency"];
        }
    }

    static fromJS(data: any): CurrencyDto {
        data = typeof data === 'object' ? data : {};
        let result = new CurrencyDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["code"] = this.code;
        data["isDefault"] = this.isDefault;
        data["factorToDefaultCurrency"] = this.factorToDefaultCurrency;
        return data;
    }
}

export interface ICurrencyDto {
    code?: string | undefined;
    isDefault?: boolean;
    factorToDefaultCurrency?: number;
}
// ...
export class FileMetadataDto implements IFileMetadataDto {
    name?: string | undefined;
    extension?: string | undefined;
    mimetype?: string | undefined;
    size?: number;
    visible?: boolean;
    uploadDate?: Date;

    constructor(data?: IFileMetadataDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.extension = _data["extension"];
            this.mimetype = _data["mimetype"];
            this.size = _data["size"];
            this.visible = _data["visible"];
            this.uploadDate = _data["uploadDate"] ? new Date(_data["uploadDate"].toString()) : <any>undefined;
        }
    }

    static fromJS(data: any): FileMetadataDto {
        data = typeof data === 'object' ? data : {};
        let result = new FileMetadataDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["extension"] = this.extension;
        data["mimetype"] = this.mimetype;
        data["size"] = this.size;
        data["visible"] = this.visible;
        data["uploadDate"] = this.uploadDate ? this.uploadDate.toISOString() : <any>undefined;
        return data;
    }
}

export interface IFileMetadataDto {
    name?: string | undefined;
    extension?: string | undefined;
    mimetype?: string | undefined;
    size?: number;
    visible?: boolean;
    uploadDate?: Date;
}

export class InvoiceDto implements IInvoiceDto {
    fileId?: string;
    fileMetadata?: FileMetadataDto;

    constructor(data?: IInvoiceDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.fileId = _data["fileId"];
            this.fileMetadata = _data["fileMetadata"] ? FileMetadataDto.fromJS(_data["fileMetadata"]) : <any>undefined;
        }
    }

    static fromJS(data: any): InvoiceDto {
        data = typeof data === 'object' ? data : {};
        let result = new InvoiceDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["fileId"] = this.fileId;
        data["fileMetadata"] = this.fileMetadata ? this.fileMetadata.toJSON() : <any>undefined;
        return data;
    }
}

export interface IInvoiceDto {
    fileId?: string;
    fileMetadata?: FileMetadataDto;
}

export class InvoiceListDto implements IInvoiceListDto {
    count?: number;
    items?: InvoiceDto[] | undefined;

    constructor(data?: IInvoiceListDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.count = _data["count"];
            if (Array.isArray(_data["items"])) {
                this.items = [] as any;
                for (let item of _data["items"])
                    this.items!.push(InvoiceDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): InvoiceListDto {
        data = typeof data === 'object' ? data : {};
        let result = new InvoiceListDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["count"] = this.count;
        if (Array.isArray(this.items)) {
            data["items"] = [];
            for (let item of this.items)
                data["items"].push(item.toJSON());
        }
        return data;
    }
}

export interface IInvoiceListDto {
    count?: number;
    items?: InvoiceDto[] | undefined;
}

export class TransactionCategoryDto implements ITransactionCategoryDto {
    id?: number;
    name?: string | undefined;
    isIncome?: boolean;

    constructor(data?: ITransactionCategoryDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.isIncome = _data["isIncome"];
        }
    }

    static fromJS(data: any): TransactionCategoryDto {
        data = typeof data === 'object' ? data : {};
        let result = new TransactionCategoryDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["isIncome"] = this.isIncome;
        return data;
    }
}

export interface ITransactionCategoryDto {
    id?: number;
    name?: string | undefined;
    isIncome?: boolean;
}
// ...
export class TransactionDto implements ITransactionDto {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    recipientName?: string | undefined;
    amount?: number;
    dateCreated?: Date;
    dateModified?: Date;
    userId?: number;
    currency?: CurrencyDto;
    transactionCategory?: TransactionCategoryDto;
    invoices?: InvoiceListDto;

    constructor(data?: ITransactionDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.title = _data["title"];
            this.description = _data["description"];
            this.recipientName = _data["recipientName"];
            this.amount = _data["amount"];
            this.dateCreated = _data["dateCreated"] ? new Date(_data["dateCreated"].toString()) : <any>undefined;
            this.dateModified = _data["dateModified"] ? new Date(_data["dateModified"].toString()) : <any>undefined;
            this.userId = _data["userId"];
            this.currency = _data["currency"] ? CurrencyDto.fromJS(_data["currency"]) : <any>undefined;
            this.transactionCategory = _data["transactionCategory"] ? TransactionCategoryDto.fromJS(_data["transactionCategory"]) : <any>undefined;
            this.invoices = _data["invoices"] ? InvoiceListDto.fromJS(_data["invoices"]) : <any>undefined;
        }
    }

    static fromJS(data: any): TransactionDto {
        data = typeof data === 'object' ? data : {};
        let result = new TransactionDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["title"] = this.title;
        data["description"] = this.description;
        data["recipientName"] = this.recipientName;
        data["amount"] = this.amount;
        data["dateCreated"] = this.dateCreated ? this.dateCreated.toISOString() : <any>undefined;
        data["dateModified"] = this.dateModified ? this.dateModified.toISOString() : <any>undefined;
        data["userId"] = this.userId;
        data["currency"] = this.currency ? this.currency.toJSON() : <any>undefined;
        data["transactionCategory"] = this.transactionCategory ? this.transactionCategory.toJSON() : <any>undefined;
        data["invoices"] = this.invoices ? this.invoices.toJSON() : <any>undefined;
        return data;
    }
}

export interface ITransactionDto {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    recipientName?: string | undefined;
    amount?: number;
    dateCreated?: Date;
    dateModified?: Date;
    userId?: number;
    currency?: CurrencyDto;
    transactionCategory?: TransactionCategoryDto;
    invoices?: InvoiceListDto;
}